Remove trait.
