+++ includes/REST_API/REST_Controller.php
@@ -30,7 +30,6 @@ namespace Google\Web_Stories\REST_API;
 use Google\Web_Stories\Infrastructure\Delayed;
 use Google\Web_Stories\Infrastructure\Registerable;
 use Google\Web_Stories\Infrastructure\Service;
-use Google\Web_Stories\Traits\Register_Routes;
 use WP_REST_Controller;
 
 /**
@@ -39,5 +38,36 @@ use WP_REST_Controller;
  * @package Google\Web_Stories\REST_API
  */
 abstract class REST_Controller extends WP_REST_Controller implements Service, Delayed, Registerable {
-	use Register_Routes;
+	/**
+	 * Register the service.
+	 *
+	 * @since 1.7.0
+	 *
+	 * @return void
+	 */
+	public function register() {
+		$this->register_routes();
+	}
+
+	/**
+	 * Get the action to use for registering the service.
+	 *
+	 * @since 1.7.0
+	 *
+	 * @return string Registration action to use.
+	 */
+	public static function get_registration_action() {
+		return 'rest_api_init';
+	}
+
+	/**
+	 * Get the action priority to use for registering the service.
+	 *
+	 * @since 1.7.0
+	 *
+	 * @return int Registration action priority to use.
+	 */
+	public static function get_registration_action_priority() {
+		return 100;
+	}
 }
