+++ includes/REST_API/Stories_Media_Controller.php
@@ -30,7 +30,6 @@ use Google\Web_Stories\Infrastructure\Delayed;
 use Google\Web_Stories\Infrastructure\Registerable;
 use Google\Web_Stories\Infrastructure\Service;
 use Google\Web_Stories\Media;
-use Google\Web_Stories\Traits\Register_Routes;
 use Google\Web_Stories\Traits\Types;
 use WP_Error;
 use WP_REST_Request;
@@ -42,7 +41,7 @@ use WP_REST_Attachments_Controller;
  */
 class Stories_Media_Controller extends WP_REST_Attachments_Controller implements Service, Delayed, Registerable {
 	use Types;
-	use Register_Routes;
+
 	/**
 	 * Constructor.
 	 *
@@ -53,6 +52,39 @@ class Stories_Media_Controller extends WP_REST_Attachments_Controller implements
 	public function __construct() {
 		parent::__construct( 'attachment' );
 		$this->namespace = 'web-stories/v1';
+	}
+
+	/**
+	 * Register the service.
+	 *
+	 * @since 1.7.0
+	 *
+	 * @return void
+	 */
+	public function register() {
+		$this->register_routes();
+	}
+
+	/**
+	 * Get the action to use for registering the service.
+	 *
+	 * @since 1.7.0
+	 *
+	 * @return string Registration action to use.
+	 */
+	public static function get_registration_action() {
+		return 'rest_api_init';
+	}
+
+	/**
+	 * Get the action priority to use for registering the service.
+	 *
+	 * @since 1.7.0
+	 *
+	 * @return int Registration action priority to use.
+	 */
+	public static function get_registration_action_priority() {
+		return 100;
 	}
 
 	/**
